moved CoordinateDistance out of GeoUtils
